+++ CHANGELOG.md
@@ -13,7 +13,8 @@
 - [add] You can also create a group that contains raw content (not a field) with `Former::group('label')->contents('YourContent')`. This will wrap the content in a control class so that your content is aligned with the fields.
 - [add] Checkables now handle being populated from relations
 - [add] You can now add classes to the group via the `->addGroupClass` method
-- [add] Former::withRules() no also take a Laravel-formatted string of rules (ie. "rule|rule:parameter|rule")
+- [add] Former::withRules() now also take a Laravel-formatted string of rules (ie. "rule|rule:parameter|rule")
+- [add] You can now populate on a form-basis with the chained method `->populate` on a form opener
 
 2.6.0
 -----